+++ brokerapi/brokers/brokers_test.go
@@ -139,7 +139,7 @@ var _ = Describe("Brokers", func() {
 
 	Describe("Broker init", func() {
 		It("should have 11 services in sevices map", func() {
-			Expect(len(gcpBroker.ServiceBrokerMap)).To(Equal(9))
+			Expect(len(gcpBroker.ServiceBrokerMap)).To(Equal(11))
 		})
 
 		It("should have a default client", func() {
@@ -153,7 +153,7 @@ var _ = Describe("Brokers", func() {
 
 	Describe("getting broker catalog", func() {
 		It("should have 11 services available", func() {
-			Expect(len(gcpBroker.Services())).To(Equal(9))
+			Expect(len(gcpBroker.Services())).To(Equal(11))
 		})
 
 		It("should have 3 storage plans available", func() {
@@ -468,12 +468,13 @@ var _ = Describe("Brokers", func() {
 var _ = Describe("AccountManagers", func() {
 
 	var (
-		logger         lager.Logger
-		iamStyleBroker models.ServiceBrokerHelper
-		spannerBroker  models.ServiceBrokerHelper
-		cloudsqlBroker models.ServiceBrokerHelper
-		accountManager modelsfakes.FakeAccountManager
-		err            error
+		logger            lager.Logger
+		iamStyleBroker    models.ServiceBrokerHelper
+		spannerBroker     models.ServiceBrokerHelper
+		cloudsqlBroker    models.ServiceBrokerHelper
+		accountManager    modelsfakes.FakeAccountManager
+		sqlAccountManager modelsfakes.FakeAccountManager
+		err               error
 	)
 
 	BeforeEach(func() {
@@ -489,7 +490,16 @@ var _ = Describe("AccountManagers", func() {
 		db_service.DbConnection = testDb
 		name_generator.New()
 
-		accountManager = modelsfakes.FakeAccountManager{}
+		accountManager = modelsfakes.FakeAccountManager{
+			CreateAccountInGoogleStub: func(instanceID string, bindingID string, details models.BindDetails, instance models.ServiceInstanceDetails) (models.ServiceBindingCredentials, error) {
+				return models.ServiceBindingCredentials{OtherDetails: "{}"}, nil
+			},
+		}
+		sqlAccountManager = modelsfakes.FakeAccountManager{
+			CreateAccountInGoogleStub: func(instanceID string, bindingID string, details models.BindDetails, instance models.ServiceInstanceDetails) (models.ServiceBindingCredentials, error) {
+				return models.ServiceBindingCredentials{OtherDetails: "{}"}, nil
+			},
+		}
 
 		iamStyleBroker = &pubsub.PubSubBroker{
 			Logger: logger,
@@ -499,8 +509,9 @@ var _ = Describe("AccountManagers", func() {
 		}
 
 		cloudsqlBroker = &cloudsql.CloudSQLBroker{
-			Logger:         logger,
-			AccountManager: &accountManager,
+			Logger:           logger,
+			AccountManager:   &sqlAccountManager,
+			SaAccountManager: &accountManager,
 		}
 
 		spannerBroker = &spanner.SpannerBroker{
@@ -542,22 +553,25 @@ var _ = Describe("AccountManagers", func() {
 
 				Expect(err).NotTo(HaveOccurred())
 				Expect(accountManager.CreateAccountInGoogleCallCount()).To(Equal(1))
+				Expect(sqlAccountManager.CreateAccountInGoogleCallCount()).To(Equal(1))
 				_, _, details, _ := accountManager.CreateAccountInGoogleArgsForCall(0)
 				Expect(details.Parameters).NotTo(BeEmpty())
 
 				username, usernameOk := details.Parameters["username"].(string)
 				password, passwordOk := details.Parameters["password"].(string)
-
+				
 				Expect(usernameOk).To(BeTrue())
 				Expect(passwordOk).To(BeTrue())
 				Expect(username).NotTo(BeEmpty())
 				Expect(password).NotTo(BeEmpty())
 			})
+
 		})
 
 		Context("when MergeCredentialsAndInstanceInfo is called on a broker", func() {
 			It("should call MergeCredentialsAndInstanceInfo on the account manager", func() {
-				_ = iamStyleBroker.BuildInstanceCredentials(make(map[string]string), make(map[string]string))
+				_, err = iamStyleBroker.BuildInstanceCredentials(models.ServiceBindingCredentials{}, models.ServiceInstanceDetails{})
+				Expect(err).ToNot(HaveOccurred())
 				Expect(accountManager.BuildInstanceCredentialsCallCount()).To(Equal(1))
 			})
 		})
@@ -577,6 +591,7 @@ var _ = Describe("AccountManagers", func() {
 				err = cloudsqlBroker.Unbind(models.ServiceBindingCredentials{})
 				Expect(err).NotTo(HaveOccurred())
 				Expect(accountManager.DeleteAccountFromGoogleCallCount()).To(Equal(1))
+				Expect(sqlAccountManager.DeleteAccountFromGoogleCallCount()).To(Equal(1))
 			})
 		})
 	})