+++ integration_tests/async_integration_test.go
@@ -153,10 +153,14 @@ var _ = Describe("AsyncIntegrationTests", func() {
 			bindDetails := models.BindDetails{
 				ServiceID: serviceNameToId[models.CloudsqlMySQLName],
 				PlanID:    serviceNameToPlanId[models.CloudsqlMySQLName],
+				Parameters: map[string]interface{}{
+					"role": "editor",
+				},
 			}
-			creds, err := gcpBroker.Bind("integration_test_instance", "binding_id", bindDetails)
+			creds, err := gcpBroker.Bind("integration_test_instance", "bind-id", bindDetails)
 			Expect(err).NotTo(HaveOccurred())
 			credsMap := creds.Credentials.(map[string]string)
+			Expect(credsMap["uri"]).To(ContainSubstring("mysql"))
 
 			// make sure we have a username and google has ssl certs
 			Expect(credsMap["Username"]).ToNot(Equal(""))
@@ -168,7 +172,7 @@ var _ = Describe("AsyncIntegrationTests", func() {
 				ServiceID: serviceNameToId[models.CloudsqlMySQLName],
 				PlanID:    serviceNameToPlanId[models.CloudsqlMySQLName],
 			}
-			err = gcpBroker.Unbind("integration_test_instance", "binding_id", unBindDetails)
+			err = gcpBroker.Unbind("integration_test_instance", "bind-id", unBindDetails)
 			Expect(err).NotTo(HaveOccurred())
 
 			// make sure google no longer has certs
@@ -200,10 +204,11 @@ var _ = Describe("AsyncIntegrationTests", func() {
 
 	Describe("Cloud SQL - PostgreSQL", func() {
 
+		var sqlService *googlecloudsql.Service
 		var dbService *googlecloudsql.InstancesService
 		var sslService *googlecloudsql.SslCertsService
 		BeforeEach(func() {
-			sqlService, err := googlecloudsql.New(gcpBroker.GCPClient)
+			sqlService, err = googlecloudsql.New(gcpBroker.GCPClient)
 			Expect(err).NotTo(HaveOccurred())
 			dbService = googlecloudsql.NewInstancesService(sqlService)
 			sslService = googlecloudsql.NewSslCertsService(sqlService)
@@ -233,8 +238,11 @@ var _ = Describe("AsyncIntegrationTests", func() {
 			bindDetails := models.BindDetails{
 				ServiceID: serviceNameToId[models.CloudsqlPostgresName],
 				PlanID:    serviceNameToPlanId[models.CloudsqlPostgresName],
+				Parameters: map[string]interface{}{
+					"role": "editor",
+				},
 			}
-			creds, err := gcpBroker.Bind("integration_test_instance", "binding_id", bindDetails)
+			creds, err := gcpBroker.Bind("integration_test_instance", "bind-id", bindDetails)
 			Expect(err).NotTo(HaveOccurred())
 			credsMap := creds.Credentials.(map[string]string)
 
@@ -242,13 +250,30 @@ var _ = Describe("AsyncIntegrationTests", func() {
 			Expect(credsMap["Username"]).ToNot(Equal(""))
 			_, err = sslService.Get(gcpBroker.RootGCPCredentials.ProjectId, cloudsqlInstanceName, credsMap["Sha1Fingerprint"]).Do()
 			Expect(err).NotTo(HaveOccurred())
+			Expect(credsMap["uri"]).To(ContainSubstring("postgres"))
+
+			// The bind oepration finishes before the instance has finished updating
+			keepWaiting := true
+			for keepWaiting == true {
+				resp, err := sqlService.Operations.List(gcpBroker.RootGCPCredentials.ProjectId, cloudsqlInstanceName).Do()
+				Expect(err).NotTo(HaveOccurred())
+				keepWaiting = false
+				for _, op := range resp.Items {
+					if op.EndTime == "" {
+						keepWaiting = true
+					}
+				}
+				// sleep for 1 second between polling so we don't hit our rate limit
+				time.Sleep(time.Second)
+			}
 
 			// unbind the instance
 			unBindDetails := models.UnbindDetails{
 				ServiceID: serviceNameToId[models.CloudsqlPostgresName],
 				PlanID:    serviceNameToPlanId[models.CloudsqlPostgresName],
 			}
-			err = gcpBroker.Unbind("integration_test_instance", "binding_id", unBindDetails)
+
+			err = gcpBroker.Unbind("integration_test_instance", "bind-id", unBindDetails)
 			Expect(err).NotTo(HaveOccurred())
 
 			// make sure google no longer has certs
@@ -260,6 +285,7 @@ var _ = Describe("AsyncIntegrationTests", func() {
 				ServiceID: serviceNameToId[models.CloudsqlPostgresName],
 				PlanID:    serviceNameToPlanId[models.CloudsqlPostgresName],
 			}
+
 			_, err = gcpBroker.Deprovision("integration_test_instance", deprovisionDetails, true)
 			Expect(err).NotTo(HaveOccurred())
 			pollForMaxFiveMins(gcpBroker, "integration_test_instance")